+++ pkg/controller/add_containersources.go
@@ -17,7 +17,7 @@ limitations under the License.
 package controller
 
 import (
-	"github.com/knative/eventing-sources/pkg/controller/containersource"
+	"github.com/knative/eventing-sources/pkg/reconciler/containersource"
 )
 
 func init() {
