+++ pkg/controller/add_githubsource.go
@@ -17,7 +17,7 @@ limitations under the License.
 package controller
 
 import (
-	"github.com/knative/eventing-sources/pkg/controller/githubsource"
+	"github.com/knative/eventing-sources/pkg/reconciler/githubsource"
 )
 
 func init() {
