+++ pkg/controller/add_kuberneteseventsource.go
@@ -17,7 +17,7 @@ limitations under the License.
 package controller
 
 import (
-	"github.com/knative/eventing-sources/pkg/controller/kuberneteseventsource"
+	"github.com/knative/eventing-sources/pkg/reconciler/kuberneteseventsource"
 )
 
 func init() {
