+++ pkg/reconciler/awssqssource/resources/receive_adapter.go
